refactor: rewrite to TS and start using Jest
