+++ .gitignore
@@ -1,2 +1,5 @@
 .vscode
 node_modules
+lib
+dist
+coverage
