+++ package.json
@@ -6,8 +6,10 @@
 		"asyncapi-converter": "cli.js"
 	},
 	"main": "lib/index.js",
+	"types": "lib/index.d.ts",
 	"scripts": {
-		"test": "mocha",
+		"build": "tsc",
+		"test": "cross-env CI=true NODE_ENV=test jest --coverage",
 		"lint": "echo 'no linter configured yet'",
 		"release": "semantic-release",
 		"generate:assets": "echo 'No additional assets need to be generated at the moment'",
@@ -38,13 +40,21 @@
 		"lodash": "^4.17.21"
 	},
 	"devDependencies": {
+		"@jest/types": "^27.5.1",
 		"@semantic-release/commit-analyzer": "^9.0.2",
 		"@semantic-release/github": "^8.0.2",
 		"@semantic-release/npm": "^8.0.3",
 		"@semantic-release/release-notes-generator": "^10.0.3",
+		"@types/jest": "^27.4.1",
+		"@types/js-yaml": "^4.0.5",
 		"conventional-changelog-conventionalcommits": "^4.6.3",
+		"cross-env": "^7.0.3",
+		"jest": "^27.5.1",
+		"mocha": "^9.1.3",
 		"semantic-release": "^18.0.1",
-		"mocha": "^9.1.3"
+		"ts-jest": "^27.1.3",
+		"ts-node": "^10.7.0",
+		"typescript": "^4.6.2"
 	},
 	"release": {
 		"branches": [
