+++ src/index.ts
@@ -1,0 +1,3 @@
+export { convert } from './convert';
+
+export type { ConvertVersion, ConvertOptions } from './interfaces';
