+++ src/interfaces.ts
@@ -1,0 +1,3 @@
+export type AsyncAPIDocument = { asyncapi: string } & Record<string, any>;
+export type ConvertVersion = string;
+export interface ConvertOptions {}
